// src/api-client/models/BoilerType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 */
export const BoilerType = {
  Opentherm: "opentherm",
  OnOff: "on_off",
  Null: "null",
} as const;
export type BoilerType = (typeof BoilerType)[keyof typeof BoilerType];

export function BoilerTypeFromJSON(json: any): BoilerType {
  return BoilerTypeFromJSONTyped(json, false);
}

export function BoilerTypeFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): BoilerType {
  return json as BoilerType;
}

export function BoilerTypeToJSON(value?: BoilerType | null): any {
  return value as any;
}