// src/api-client/models/MaxSoundLevel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * The maximum sound level selected
 * @export
 */
export const MaxSoundLevel = {
  Normal: "normal",
  Library: "library",
  Silent: "silent",
  Building87: "building87",
} as const;
export type MaxSoundLevel = (typeof MaxSoundLevel)[keyof typeof MaxSoundLevel];

export function MaxSoundLevelFromJSON(json: any): MaxSoundLevel {
  return MaxSoundLevelFromJSONTyped(json, false);
}

export function MaxSoundLevelFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): MaxSoundLevel {
  return json as MaxSoundLevel;
}

export function MaxSoundLevelToJSON(value?: MaxSoundLevel | null): any {
  return value as any;
}