// src/api-client/models/SilentMode.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 */
export const SilentMode = {
  Never: "never",
  Night: "night",
  Always: "always",
} as const;
export type SilentMode = (typeof SilentMode)[keyof typeof SilentMode];

export function SilentModeFromJSON(json: any): SilentMode {
  return SilentModeFromJSONTyped(json, false);
}

export function SilentModeFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): SilentMode {
  return json as SilentMode;
}

export function SilentModeToJSON(value?: SilentMode | null): any {
  return value as any;
}